Add a method for defining the module cache key
Adds a method "getCacheKey()" on Aether Modules. The method acts as a
way to hook in to the creation of the cache key for a module. It allows
for overriding or appending the key with, for example, a key
representating a given state.

Note that the method accepts an argument "$key". This key is the
original cache key. It is recommended to append a value to it instead
of completely overriding it.